pyproject.toml
[tool.poetry]
name = "django"
version = "4.2.1"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "^3.7.2"
async-timeout = "^4.0.3"
billiard = "^4.1.0"
celery = "^5.3.1"
charset-normalizer = "^3.2.0"
click = "^8.1.6"
click-didyoumean = "^0.3.0"
click-plugins = "^1.1.1"
django-celery-results = "^2.5.1"
gunicorn = "^21.2.0"
idna = "^3.4"
packaging = "^23.1"
pillow = "^10.0.0"
python-dateutil = "^2.8.2"
redis = "^5.0.0"
requests = "^2.31.0"
six = "^1.16.0"
sqlparse = "^0.4.4"
tzdata = "^2023.3"
whitenoise = "^6.5.0"
pymysql = "^1.1.0"
django-dotenv = "^1.4.2"
social-auth-app-django = "^5.2.0"
pyvimeo = "^1.1.0"
django-browser-reload = "^1.11.0"
django-redis = "^5.3.0"
psycopg = "^3.1.10"
psycopg2 = "^2.9.7"
django-heroku = "^0.3.1"
urllib3 = "^2.0.4"
kombu = "^5.3.2"
dj-database-url = "^2.1.0"
python-decouple = "^3.8"
django-debug-toolbar = "^4.2.0"
mailgun2 = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
